flask-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
redis==5.0.1
python-socketio==5.9.0
eventlet==0.33.3
gunicorn==21.2.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.0.0.post2
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
six==1.16.0
PyJWT==2.8.0
blinker==1.6.2
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
dnspython==2.4.2
Flask-Mail==0.10.0
Flask-CORS==4.0.0